refactor(registry/crd): address review feedback
- rename package registry -> crd to match its directory and drop the
  now-redundant import alias in the factory
- remove dead IsEndpoint() and EndpointLabels() methods
- fix DNSRecordSpec doc comment (DNSEndpoint -> DNSRecord) and drop the
  redundant +groupName marker already declared at package level
- rename the singular field's JSON tag endpoints -> endpoint and
  regenerate the CRD manifests
- exercise a distinct UpdateOld endpoint in the ApplyChanges update test
